helm-generator-service/requirements.txt
attrs==25.3.0
click==8.2.1
colorama==0.4.6
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
tomli==2.2.1
types-jsonschema==4.25.1.20250821
typing_extensions==4.14.1
